rename: createTenantValidator to tenantValidator
