+++ src/validators/index.ts
@@ -3,4 +3,4 @@ export { default as verifyOtpValidator } from "./register/verifyOtpValidator";
 export { default as forgetPasswordValidator } from "./forgetPasswordValidator";
 export { default as loginValidator } from "./loginValidator";
 export { default as setPasswordValidator } from "./setPasswordValidator";
-export { default as createTenantValidator } from "./createTenantValidator";
+export { default as tenantValidator } from "./tenantValidator";
